+++ core/designsystem/src/commonMain/kotlin/io/spherelabs/designsystem/pininput/LKPinInput.kt
@@ -1,0 +1,80 @@
+package io.spherelabs.designsystem.pininput
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalSoftwareKeyboardController
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalComposeUiApi::class)
+@Composable
+fun LKPinInput(
+    modifier: Modifier = Modifier,
+    cellModifier: Modifier = Modifier,
+    style: LKPinStyle = LKPinDefaults.style(),
+    value: String = "",
+    disableKeypad: Boolean = false,
+    onValueChanged: (String) -> Unit
+) {
+    val focusRequester = remember { FocusRequester() }
+    val keyboard = LocalSoftwareKeyboardController.current
+
+    val length by style.cellCount()
+
+    TextField(
+        readOnly = disableKeypad,
+        value = value,
+        onValueChange = {
+            if (it.length <= length) {
+                if (it.all { c -> c in '0'..'9' }) {
+                    onValueChanged(it)
+                }
+                if (it.length >= length) {
+                    keyboard?.hide()
+                }
+            }
+        },
+        modifier = Modifier
+            .size(0.dp)
+            .focusRequester(focusRequester),
+        keyboardOptions = KeyboardOptions(
+            keyboardType = KeyboardType.Number
+        )
+    )
+
+    Row(
+        modifier = modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.Center
+    ) {
+        repeat(length) {
+            LKCell(
+                modifier = cellModifier
+                    .size(width = 75.dp, height = 75.dp)
+                    .clip(RoundedCornerShape(24.dp))
+                    .background(color = Color.White)
+                    .clickable {
+                        focusRequester.requestFocus()
+                        keyboard?.show()
+                    },
+                value = value.getOrNull(it),
+                isCursorVisible = value.length == it,
+                style.obscureText().value
+            )
+            if (it != length - 1)
+                Spacer(modifier = Modifier.size(8.dp))
+        }
+    }
+}
